impl affect_rows in rest_api.rs and use query_iter
